+++ client/src/pages/register/Register.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { Bodytext, Heading2 } from '../../shared-styles';
+import { Bodytext, Heading2 } from '../../styles/shared-styles';
 
 /**
  *TODO: the form for registering, can be accessed from the login page
