+++ client/src/styles/variables.ts
